--- docker/App.Dockerfile
@@ -1,11 +1,0 @@
-FROM golang:1.23
-
-WORKDIR /app
-
-COPY go.mod go.sum ./
-RUN go mod download && go mod verify
-
-COPY . .
-RUN CGO_ENABLED=0 GOOS=linux go build -v -o /app/moddingway-app
-
-ENTRYPOINT ["/app/moddingway-app"]
